I am adding files
Please be aware that I changes some of the code and added bubbles. I also used R to calculate the multivariate regressions for the data. R^2 is about 89% so we are looking good. Look at the data.analysis.notes.txt for the full analysis.

Also, the bubbles are nowhere near done. I need to add a color scale, my tooltip sucks, and I haven't figured out how to add levels to the bubbles so I will need some help with those items.

--Steven